+++ lib/alchem_iiif_web.ex
@@ -88,8 +88,8 @@ defmodule AlchemIiifWeb do
       import AlchemIiifWeb.CoreComponents
 
       # Common modules used in templates
-      alias Phoenix.LiveView.JS
       alias AlchemIiifWeb.Layouts
+      alias Phoenix.LiveView.JS
 
       # Routes generation with the ~p sigil
       unquote(verified_routes())